[INLONG-8171][Manager] Optimize the way of creating LoadNodes to make it easy to expand and maintain
